\name{circFGN}
\alias{circFGN}

\title{Simulation of a fractional Gaussian noise}
\description{
  Generates a sample path of fractional Gaussian noise with scaling coefficient \code{C=1} and Hurst parameter \code{H} in (0,1)
}
\usage{
circFGN(n, H, C=1, plot=FALSE)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{n}{ sample size }
  \item{H}{ Hurst parameter }
  \item{C}{ scaling coefficient }
  \item{plot}{plot of the generated sample path}
}
\details{
This simulation is achieved by computing the lag difference of a fractional Brownian motion which is generated by the method in the package \code{longmemo}.
}
\value{
  a vector of length \code{n}.
}
\references{ 
Jan Beran (1994) Statistics for Long-Memory Processes, Chapman & Hall.
}

\author{
Jan Beran (original) and Wonsang You (fine tuning)
}

\seealso{ \code{\link{perturbFGN}}, \code{\link{wFGN}} }
\examples{
n<-10000; H<-0.2
ts <- circFGN(n,H)
}